增加上传文件限制

+++ CMS/cms/service/notice.py
@@ -17,32 +17,41 @@ class NoticeService(object):
 
     upload_uri = '/res/img/upload?scene=common'
     def request_upload(self, pic):
-        file_type = pic.filename.split('.')[1]
-        file_name = '.'.join([str(random.randint(6, 1000000)), file_type])
-        file_url = os.path.join(
-            app.config['TEMP_DIR'], file_name
-        )
-        pic.save(file_url)
+        if pic and self.allow_file(pic.filename):
+            file_type = pic.filename.rsplit('.', 1)[1]
+            pic.filename = '{0}.{1}'.format(
+                str(random.randint(6, 1000000)), file_type
+            )
+            file_url = os.path.join(
+                app.config['TEMP_DIR'], file_type
+            )
+            # 保存图片
+            pic.save(file_url)
 
-        register_openers()
-        datagen, headers = multipart_encode({"fileUp": open(file_url, 'rb')})
+            register_openers()
+            datagen, headers = multipart_encode({"fileUp": open(file_url, 'rb')})
 
-        url = '{0}{1}'.format(app.config['SERVICE_URL'], self.upload_uri)
-        request = urllib2.Request(
-            url,
-            datagen, headers)
-        # 删除图片
-        os.remove(file_url)
-        resp = json.loads(urllib2.urlopen(request).read())
+            url = '{0}{1}'.format(app.config['SERVICE_URL'], self.upload_uri)
+            request = urllib2.Request(
+                url,
+                datagen, headers)
+            # 删除图片
+            os.remove(file_url)
+            resp = json.loads(urllib2.urlopen(request).read())
 
-        img_id = resp['content']['id']
-        img_url = resp['content']['origin']
+            img_id = resp['content']['id']
+            img_url = resp['content']['origin']
 
-        self.save_data(img_id, img_url)
+            self.save_data(img_id, img_url)
 
-        if resp['ret'] == 0:
-            return resp['content'], resp['timestamp']
+            if resp['ret'] == 0:
+                return resp['content'], resp['timestamp']
+            return None
         return None
+
+    def allow_file(self, filename):
+        return '.' in filename and filename.rsplit('.', 1)[1] \
+                                   in app.config['ALLOWED_EXTENSIONS']
 
     def save_data(self, img_id, img_url, img_name=None):
         notice = NoticeModel(
